+++ core/src/elements/ons-dialog/index.spec.js
@@ -67,11 +67,11 @@ describe('OnsDialogElement', () => {
     });
   });
 
-  describe('#_onDeviceBackButton()', () => {
+  describe('#onDeviceBackButton', () => {
     it('cancels if dialog is cancelable', () => {
       const spy = chai.spy.on(dialog, '_cancel');
       dialog.setAttribute('cancelable', '');
-      dialog._onDeviceBackButton();
+      dialog.onDeviceBackButton._callback();
       expect(spy).to.have.been.called.once;
     });
 
@@ -79,7 +79,7 @@ describe('OnsDialogElement', () => {
       const event = {};
       const spy = chai.spy.on(event, 'callParentHandler');
 
-      dialog._onDeviceBackButton(event);
+      dialog.onDeviceBackButton._callback(event);
       expect(spy).to.have.been.called.once;
     });
   });