+++ CHANGELOG.md
@@ -3,8 +3,13 @@ CHANGELOG
 ====
 
 v2.0.0-rc.7
+----
  * ons-carousel: Support ng-repeat in Angular1 bindings.
  * ons-carousel: Fix [#1168](https://github.com/OnsenUI/OnsenUI/issues/1168).
+ * ons-switch: Trigger `change` event only once.
+ * ons-bottom-toolbar: Correctly register element. Fix [#1426](https://github.com/OnsenUI/OnsenUI/issues/1426).
+ * ons-pull-hook: Fix [#1415](https://github.com/OnsenUI/OnsenUI/issues/1415).
+ * ons-navigator: Add `onDeviceBackButton` property.
 
 v2.0.0-rc.6
 ----
@@ -14,6 +19,7 @@ v2.0.0-rc.6
  * ons-speed-dial, ons-speed-dial-item: Removed `_compiled` attribute.
  * ons-speed-dial: Make ons-speed-dial elements angular2 compatible.
  * ons-carousel: Fix [#1423](https://github.com/OnsenUI/OnsenUI/issues/1423).
+ * core: Create unified `onDeviceBackButton` API for alert-dialog, dialog, modal, page, popover, splitter.
 
 v2.0.0-rc.5
 ----